requirements.txt
aiosmtplib==4.0.2
annotated-types==0.7.0
anyio==4.11.0
babel==2.17.0
backrefs==5.9
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fakeredis==2.32.0
fastapi==0.119.1
fastapi-mail==1.5.7
flake8==7.3.0
ghp-import==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
jiter==0.11.1
Markdown==3.9
MarkupSafe==3.0.3
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.22
mkdocs-material-extensions==1.3.1
openai==2.6.0
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.5.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
pymdown-extensions==10.16.1
pypdf==6.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
pyyaml_env_tag==1.1
redis==6.4.0
regex==2025.10.22
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
